fix: Register web additional-beans discoverer under the declaring namespace
Verifying the MockMvc part of this PR by disabling it revealed that the new
highlighting tests passed with the fix removed - they never exercised it. A probe
dumping `SpringSearchService.getStaticBeans(module)` returned only the five
spring-core beans, with no spring-web bean at all.

Root cause: `spring-web-plugin.xml` registered `SpringWebAdditionalBeansDiscoverer`
under `defaultExtensionNs="com.explyt.spring.web"`, while the extension point is
declared by spring-core as `com.explyt.spring.core.additionalBeansDiscoverer`. The
registration therefore matched no extension point and the discoverer was never
invoked, so `WebApplicationContext`, the servlet objects and the newly added
`MockMvc` were all missing from the bean model. `spring-data` and
`spring-integration` already use the correct namespace.

Replaces the two ineffective highlighting tests with a service-level test that
asserts what the discoverer actually contributes. Red-green verified: with the old
namespace it fails listing the five spring-core beans, with the fix it passes.
