+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - fix: Do not report `@Autowired` members of `@ContextConfiguration` test classes as not being a Spring bean
 
 ### Spring Web
+- fix: Register the web additional-beans discoverer under the extension namespace that declares the extension point, so framework-provided web beans such as `WebApplicationContext` are resolved again
 - fix: Resolve `MockMvc` autowired in tests as a bean provided by test auto-configuration
 
 ## 262.34.101 - 2026-08-18
